requirements.txt
flask 
flask-sqlalchemy 
flask-wtf
wtforms
flask-testing
pytest
pytest-cov